% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsVL.R
\name{embedSourceVL}
\alias{embedSourceVL}
\title{Creates an hyperlink to resources used in the Rmd file}
\usage{
embedSourceVL(dirs = ".", exts = c("Rmd", "R", "txt", "dat", "csv", "xls",
  "xlsx", "gz"), otherF = NULL, textT = "Source")
}
\arguments{
\item{dirs}{Path to the file}

\item{exts}{extensions of the files being included in the archive}

\item{otherF}{vector of other files to include in the archive}

\item{textT}{(optional) text for the hyperlink}
}
\value{
A string that can be printed in an html document
}
\description{
Creates an hyperlink to resources used in the Rmd file
}

